requirements.txt
## The following requirements were added by pip freeze:
autopep8==1.5
certifi==2019.11.28
chardet==3.0.4
Click==7.0
dnspython==1.16.0
eventlet==0.25.1
Flask==1.1.1
greenlet==0.4.15
gunicorn==20.0.4
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
monotonic==1.5
netifaces==0.10.6
psycopg2==2.8.4
pycodestyle==2.5.0
pycryptodome==3.9.7
python-dotenv==0.10.5
python-engineio==3.12.1
python-ffmpeg-video-streaming==0.0.25
python-socketio==4.5.1
requests==2.23.0
six==1.14.0
socketio==0.2.1
urllib3==1.25.8
virtualenv==16.7.8
websocket-client==0.57.0
Werkzeug==0.16.1
